requirements.txt
arxiv>=2.1.3
requests>=2.31.0
pandas>=2.2.0
PyMuPDF>=1.24.0
pydantic>=2.7.0
networkx>=3.3
pyvis>=0.3.2
streamlit>=1.35.0
python-dotenv>=1.0.1
google-generativeai>=0.7.0
openai>=1.35.0
neo4j>=5.0.0
tqdm>=4.66.0
pytest>=8.0.0
PyYAML>=6.0.0
ruff>=0.13.0
black>=25.1.0
pre-commit>=4.0.0
